requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
websockets==15.0.1
pymysql==1.1.1
python-dotenv==1.0.1